pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "torch-align-med"
version = "0.1.0"
description = "Semantic-alignment metrics between latent spaces. The alignment-evaluation analog of torch-fidelity."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "torch-align-med contributors" }]
keywords = ["pytorch", "alignment", "metrics", "diffusion", "representation-learning", "medical-imaging"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
encoders-2d = ["transformers>=4.40", "timm>=1.0", "open_clip_torch>=2.24"]
encoders-medical = ["transformers>=4.40", "open_clip_torch>=2.24"]
encoders-3d = ["monai>=1.3", "nibabel>=5.0"]
srss = ["segment-anything-2 ; python_version>='3.10'"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]
all = [
    "torch-align-med[encoders-2d,encoders-medical,encoders-3d]",
]

[project.scripts]
torch-align-med = "torch_align_med.cli:main"

[project.urls]
Homepage = "https://github.com/AICONSlab/torch-align-med"
Documentation = "https://github.com/AICONSlab/torch-align-med"
Repository = "https://github.com/AICONSlab/torch-align-med"

[tool.hatch.build.targets.wheel]
packages = ["src/torch_align_med"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
